@@ -1,18 +1,11 @@
-class User < ActiveRecord::Base
-  # write associations here
-    enum role: [:admin]
+class User < ApplicationRecord
+  has_secure_password
+  has_many :rides
+  has_many :attractions, through: :rides
 
-    has_secure_password
-
-  	has_many :rides
-  	has_many :attractions, :through => :rides
-
-  	def mood
-  		if self.nausea > self.happiness
-  			"sad"
-  		else
-  			"happy"
-  		end
- 	 end
-
+  def mood
+    unless admin
+      happiness > nausea ? 'happy' : 'sad'
+    end
+  end
 end
