@@ -1,24 +1,18 @@
 class ApplicationController < ActionController::Base
-  # Prevent CSRF attacks by raising an exception.
-  # For APIs, you may want to use :null_session instead.
   protect_from_forgery with: :exception
-
-  before_action :current_user
-  # before_action :require_logged_in, except: [:new, :create, :home]
-
+  before_action :verify_user_is_authenticated
   helper_method :current_user
 
-  def logged_in?
+  private
+  def verify_user_is_authenticated
+    redirect_to '/' unless user_is_authenticated
+  end
+
+  def user_is_authenticated
     !!current_user
   end
 
-  private
-    def require_logged_in
-      redirect_to '/' unless logged_in?
-    end
-
-    def current_user
-      @current_user ||= User.find(session[:user_id]) if session[:user_id]
-    end
-
+  def current_user
+    User.find_by(id:session[:user_id])
+  end
 end
