@@ -1,4 +1,3 @@
-# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -11,10 +10,10 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20140703140419) do
+ActiveRecord::Schema.define(version: 2021_01_07_211555) do
 
   create_table "attractions", force: :cascade do |t|
-    t.string  "name"
+    t.string "name"
     t.integer "tickets"
     t.integer "nausea_rating"
     t.integer "happiness_rating"
@@ -24,16 +23,20 @@ ActiveRecord::Schema.define(version: 20140703140419) do
   create_table "rides", force: :cascade do |t|
     t.integer "user_id"
     t.integer "attraction_id"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
   end
 
   create_table "users", force: :cascade do |t|
-    t.string  "name"
-    t.string  "password_digest"
+    t.string "name"
+    t.string "password_digest"
     t.integer "nausea"
     t.integer "happiness"
     t.integer "tickets"
     t.integer "height"
-    t.boolean "admin",           default: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.boolean "admin", default: false
   end
 
 end
